Unobtrusive login
#IDEA-221 Fixed
#IDEA-223 Fixed
#IDEA-224 Fixed

+++ stepik-java-api/src/main/java/org/stepik/api/queries/metrics/StepikMetricsPostQuery.java
@@ -26,7 +26,7 @@ public class StepikMetricsPostQuery extends StepikAbstractPostQuery<VoidResult> 
     }
 
     @NotNull
-    public StepikMetricsPostQuery timestamp(int value) {
+    public StepikMetricsPostQuery timestamp(long value) {
         metrics.getMetric().setTimestamp(value);
         return this;
     }
